@@ -4,13 +4,13 @@
   "private": true,
   "desciption": "Pokédot web client",
   "scripts": {
-    "build": "next build && next export -o build",
-    "build-storybook": "build-storybook",
+    "build": "next build",
+    "build:sb": "storybook build",
     "dev": "next dev",
+    "dev:sb": "storybook dev -p 6006 --no-open",
     "lint": "eslint --cache --max-warnings 0 .",
     "lint:fix": "eslint --cache --max-warnings 0 --fix .",
-    "start": "NODE_ENV=production node serve.js",
-    "storybook": "start-storybook -p 6006"
+    "start": "NODE_ENV=production node serve.js"
   },
   "repository": {
     "type": "git",
@@ -28,33 +28,38 @@
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.58",
-    "@material-ui/styles": "^4.11.4",
+    "@emotion/react": "^11.11.4",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
     "@next/bundle-analyzer": "^10.2.2",
-    "@pkmn/data": "^0.7.10",
-    "@pkmn/dex": "^0.7.10",
-    "@pkmn/img": "^0.2.2",
-    "@pkmn/sets": "^5.0.1",
-    "@pkmn/sim": "^0.7.10",
-    "@storybook/addon-actions": "^6.2.9",
-    "@storybook/addon-essentials": "^6.2.9",
-    "@storybook/addon-links": "^6.2.9",
-    "@storybook/react": "^6.2.9",
-    "@storybook/theming": "^6.2.9",
+    "@pkmn/data": "^0.9.0",
+    "@pkmn/dex": "^0.9.0",
+    "@pkmn/img": "^0.2.25",
+    "@pkmn/sets": "^5.1.2",
+    "@pkmn/sim": "^0.9.0",
+    "@storybook/addon-actions": "^8.0.5",
+    "@storybook/addon-essentials": "^8.0.5",
+    "@storybook/addon-links": "^8.0.5",
+    "@storybook/addon-themes": "^8.0.5",
+    "@storybook/manager-api": "^8.0.5",
+    "@storybook/nextjs": "^8.0.5",
+    "@storybook/preview-api": "^8.0.5",
+    "@storybook/react": "^8.0.5",
+    "@storybook/theming": "^8.0.5",
     "@types/highlight-words-core": "^1.2.0",
     "@types/lodash": "^4.14.170",
     "@types/material-ui": "^0.21.8",
-    "@types/node": "^15.6.0",
-    "@types/react": "^17.0.6",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.73",
     "@types/react-virtualized": "^9.21.11",
-    "@typescript-eslint/eslint-plugin": "^4.24.0",
-    "@typescript-eslint/parser": "^4.24.0",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
     "axios": "^0.21.1",
     "babel-loader": "^8.2.2",
     "colyseus.js": "^0.14.12",
-    "eslint": "^7.27.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.3",
     "eslint-plugin-prettier": "^3.4.0",
@@ -63,13 +68,14 @@
     "highlight-words-core": "^1.2.2",
     "immer": "^9.0.2",
     "lodash": "^4.17.21",
-    "next": "^10.2.2",
+    "next": "^14.1.4",
     "prettier": "^2.3.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-virtualized": "^9.22.3",
-    "smogon": "^1.4.5",
-    "typescript": "^4.2.4",
+    "smogon": "^3.1.3",
+    "storybook": "^8.0.5",
+    "typescript": "^5.4.2",
     "workbox-webpack-plugin": "^6.1.5",
     "ws": "^7.4.5"
   }
