@@ -25,8 +25,8 @@
   "dependencies": {
     "@colyseus/command": "^0.1.6",
     "@colyseus/monitor": "^0.14.3",
-    "@pkmn/data": "^0.7.10",
-    "@pkmn/dex": "^0.7.10",
+    "@pkmn/data": "^0.9.0",
+    "@pkmn/dex": "^0.9.0",
     "axios": "^0.21.1",
     "colyseus": "^0.14.13",
     "config": "^3.3.6",
@@ -42,7 +42,7 @@
     "pino-debug": "^2.0.0",
     "reflect-metadata": "^0.1.13",
     "request": "^2.88.2",
-    "smogon": "^1.4.5",
+    "smogon": "^3.1.3",
     "typescript-ioc": "^3.2.2"
   },
   "devDependencies": {
@@ -50,26 +50,26 @@
     "@types/cors": "^2.8.10",
     "@types/express": "^4.17.11",
     "@types/express-pino-logger": "^4.0.2",
-    "@types/jest": "^26.0.23",
+    "@types/jest": "^29.5.12",
     "@types/lodash": "^4.14.170",
-    "@types/node": "^15.6.0",
+    "@types/node": "^20.11.30",
     "@types/pino": "^6.3.8",
     "@types/request": "^2.48.5",
-    "@typescript-eslint/eslint-plugin": "^4.24.0",
-    "@typescript-eslint/parser": "^4.24.0",
-    "eslint": "^7.27.0",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.23.3",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "jest": "^26.6.3",
+    "jest": "^29.7.0",
     "nodemon": "^2.0.7",
     "pino-pretty": "^4.8.0",
     "prettier": "^2.3.0",
     "source-map-support": "^0.5.19",
-    "ts-jest": "^26.5.6",
+    "ts-jest": "^29.1.2",
     "ts-mockito": "^2.6.1",
     "ts-node": "^9.1.1",
-    "typescript": "^4.2.4"
+    "typescript": "^5.4.2"
   }
 }
