@@ -1,5 +1,4 @@
-import {Popper, TextField, Typography} from '@material-ui/core';
-import {Autocomplete} from '@material-ui/lab';
+import {Autocomplete, Box, Popper, TextField, Typography} from '@mui/material';
 import {PokemonSet} from '@pkmn/data';
 import React, {useContext, useEffect, useState} from 'react';
 
@@ -33,7 +32,7 @@ export const PokemonAbilityInput: React.FC<PokemonAbilityInputProps> = ({
         0
       );
       const abilities = Object.fromEntries<number>(
-        Object.values(specie.abilities).map(name => {
+        (Object.values(specie.abilities) as string[]).map(name => {
           const ability = generation.abilities.get(name);
           const usage = (ability && abilityStats[ability.id]) ?? 0;
 
@@ -57,9 +56,11 @@ export const PokemonAbilityInput: React.FC<PokemonAbilityInputProps> = ({
           placeholder="Select an ability"
           size="small"
           fullWidth
+          variant={'standard'}
+          margin={'dense'}
         />
       )}
-      renderOption={option => {
+      renderOption={(props, option) => {
         let text = option;
         const percent = abilities[option];
 
@@ -67,7 +68,11 @@ export const PokemonAbilityInput: React.FC<PokemonAbilityInputProps> = ({
           text += ` (${percent.toFixed(2)}%)`;
         }
 
-        return <Typography noWrap>{text}</Typography>;
+        return (
+          <Box component={'li'} {...props}>
+            <Typography noWrap>{text}</Typography>
+          </Box>
+        );
       }}
       onChange={(_, value) => {
         onChange(pokemon => {
