@@ -6,7 +6,7 @@ import {
   DialogContentText,
   DialogTitle,
   Typography,
-} from '@material-ui/core';
+} from '@mui/material';
 import React from 'react';
 
 export interface TeamValidationDialogProps {
