@@ -1,12 +1,12 @@
 import {
+  Alert,
   Button,
   CircularProgress,
   Grid,
   Slider,
   Switch,
   Typography,
-} from '@material-ui/core';
-import {Alert} from '@material-ui/lab';
+} from '@mui/material';
 import {produce} from 'immer';
 import {useRouter} from 'next/router';
 import React, {useContext, useState} from 'react';
@@ -35,6 +35,8 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
   const handleCreateRoom = () => {
     createRoom('team-generator', options, room => {
       return router.push(`/team-generator/${room.id}`);
+    }).catch(e => {
+      console.log('Failed to create room.', e);
     });
   };
 
@@ -50,10 +52,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
     );
   };
 
-  const handlePoolsChange = (
-    _: React.ChangeEvent<unknown>,
-    value: number | number[]
-  ) => {
+  const handlePoolsChange = (_: Event, value: number | number[]) => {
     setOptions(
       produce(options, draft => {
         const indices = Array.from(Array(value as number).keys());
@@ -80,12 +79,12 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
       {!editPools && (
         <Grid
           container
-          justify="center"
+          justifyContent="center"
           alignItems="center"
           style={{height: '100%'}}
         >
-          <Grid item container justify="center" spacing={1}>
-            <Grid item container justify="center">
+          <Grid item container justifyContent="center" spacing={1}>
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Grid container spacing={2} alignItems="center">
                   <Grid item xs>
@@ -112,7 +111,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
               </Grid>
             </Grid>
 
-            <Grid item container justify="center">
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Grid container spacing={2} alignItems="center">
                   <Grid item xs>
@@ -144,7 +143,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
               </Grid>
             </Grid>
 
-            <Grid item container justify="center">
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Grid container spacing={2} alignItems="center">
                   <Grid item xs>
@@ -181,7 +180,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
               </Grid>
             </Grid>
 
-            <Grid item container justify="center">
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Grid container spacing={2} alignItems="center">
                   <Grid item xs>
@@ -214,7 +213,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
             </Grid>
 
             {error && (
-              <Grid item container justify="center">
+              <Grid item container justifyContent="center">
                 <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                   <Alert variant="outlined" severity="error">
                     {error}
@@ -224,14 +223,14 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
             )}
 
             {isLoading && (
-              <Grid item container justify="center">
+              <Grid item container justifyContent="center">
                 <Grid item>
                   <CircularProgress size={24} />
                 </Grid>
               </Grid>
             )}
 
-            <Grid item container justify="center">
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Button
                   onClick={() => setEditPools(true)}
@@ -244,7 +243,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
               </Grid>
             </Grid>
 
-            <Grid item container justify="center">
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Button
                   onClick={handleCreateRoom}
@@ -258,7 +257,7 @@ export const CreateRoom: React.FC<CreateRoomProps> = ({
               </Grid>
             </Grid>
 
-            <Grid item container justify="center">
+            <Grid item container justifyContent="center">
               <Grid item xs={12} sm={6} md={4} lg={3} xl={2}>
                 <Button
                   onClick={onBack}
