@@ -10,8 +10,7 @@ import {
   ListItemAvatar,
   ListItemText,
   Typography,
-  makeStyles,
-} from '@material-ui/core';
+} from '@mui/material';
 import React, {useContext} from 'react';
 import {AutoSizer, List, ListRowProps} from 'react-virtualized';
 
@@ -19,12 +18,6 @@ import {SpeciesImage, SpeciesImageType} from '../species/SpeciesImage';
 import {SpeciesContext} from '../species/SpeciesProvider';
 
 import {Pokemon} from './TeamGeneratorState';
-
-const useStyles = makeStyles({
-  list: {
-    outline: 'none',
-  },
-});
 
 export interface EligiblePokemonDialogProps {
   open: boolean;
@@ -41,7 +34,6 @@ export const EligiblePokemonDialog: React.FC<EligiblePokemonDialogProps> = ({
   error,
   eligiblePokemon,
 }: EligiblePokemonDialogProps) => {
-  const classes = useStyles();
   const {setOpen, setSpecies} = useContext(SpeciesContext);
   const sortedEligiblePokemon = eligiblePokemon.sort(
     (left, right) => left.num - right.num
@@ -81,7 +73,7 @@ export const EligiblePokemonDialog: React.FC<EligiblePokemonDialogProps> = ({
 
       <DialogContent>
         {isLoading && (
-          <Grid container justify="center">
+          <Grid container justifyContent="center">
             <Grid item>
               <CircularProgress />
             </Grid>
@@ -89,7 +81,7 @@ export const EligiblePokemonDialog: React.FC<EligiblePokemonDialogProps> = ({
         )}
 
         {!isLoading && error && (
-          <Grid container justify="center">
+          <Grid container justifyContent="center">
             <Grid item>
               <Typography>{error}</Typography>
             </Grid>
@@ -101,12 +93,14 @@ export const EligiblePokemonDialog: React.FC<EligiblePokemonDialogProps> = ({
             <AutoSizer>
               {({height, width}) => (
                 <List
-                  className={classes.list}
                   height={height}
                   width={width}
                   rowHeight={46}
                   rowCount={sortedEligiblePokemon.length}
                   rowRenderer={renderRow}
+                  sx={{
+                    outline: 'none',
+                  }}
                 />
               )}
             </AutoSizer>
