@@ -5,7 +5,7 @@ import {
   ListItemText,
   ListSubheader,
   Paper,
-} from '@material-ui/core';
+} from '@mui/material';
 import React, {useContext} from 'react';
 
 import {TeamGeneratorContext} from '../TeamGeneratorProvider';
