@@ -1,4 +1,4 @@
-import {CircularProgress, Grid, Typography} from '@material-ui/core';
+import {CircularProgress, Grid, Typography} from '@mui/material';
 import dynamic from 'next/dynamic';
 import {useRouter} from 'next/router';
 import React from 'react';
@@ -6,19 +6,17 @@ import React from 'react';
 import {useJoinRoom, useRoomListeners} from '../../../hooks/useRoom';
 import {TeamGeneratorContext} from '../TeamGeneratorProvider';
 
-const SetPlayerName = dynamic<unknown>(() =>
+const SetPlayerName = dynamic<object>(() =>
   import('./SetPlayerName').then(m => m.SetPlayerName)
 );
 
-const SetPlayerReady = dynamic<unknown>(() =>
+const SetPlayerReady = dynamic<object>(() =>
   import('./SetPlayerReady').then(m => m.SetPlayerReady)
 );
 
-const Pool = dynamic<unknown>(() => import('./Pool').then(m => m.Pool));
+const Pool = dynamic<object>(() => import('./Pool').then(m => m.Pool));
 
-const Summary = dynamic<unknown>(() =>
-  import('./Summary').then(m => m.Summary)
-);
+const Summary = dynamic<object>(() => import('./Summary').then(m => m.Summary));
 
 export const RoomPage: React.FC = () => {
   const {query} = useRouter();
@@ -37,7 +35,7 @@ export const RoomPage: React.FC = () => {
     content = (
       <Grid
         container
-        justify="center"
+        justifyContent="center"
         alignContent="center"
         style={{height: '100%'}}
       >
