@@ -1,4 +1,4 @@
-import {Story} from '@storybook/react';
+import {StoryFn} from '@storybook/react';
 import React from 'react';
 
 import {SummaryCard, SummaryCardProps} from '../modules/coverage/SummaryCard';
@@ -37,7 +37,7 @@ export default {
   },
 };
 
-const TheSummaryCard: Story<SummaryCardProps> = args => {
+const TheSummaryCard: StoryFn<SummaryCardProps> = args => {
   return <SummaryCard {...args} />;
 };
 
